Mage: Changed Horrible IO Expander to Physical Position System
Previously, there was an individual if statement for every possible key
and a corrisponding secondary check to see if there is room in the
buffer. Now there is an array and a loop. That's it. Why didn't I do
this in the first place it was so easy SMH. Anyways the file is now 200
lines instead of 700 so thats good.
Also, I fixed a bug that make the keyboard ignore the first key. (Using
zero as a default/unset value is great am I right).